satellite/payments: project usage charges (#3512)
