+++ satellite/console/consoleweb/consoleql/project.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/graphql-go/graphql"
 
+	"storj.io/storj/satellite/accounting"
 	"storj.io/storj/satellite/console"
 )
 
@@ -362,7 +363,7 @@ func fromMapProjectInfo(args map[string]interface{}) (project console.ProjectInf
 }
 
 // fromMapBucketUsageCursor creates console.BucketUsageCursor from input args
-func fromMapBucketUsageCursor(args map[string]interface{}) (cursor console.BucketUsageCursor) {
+func fromMapBucketUsageCursor(args map[string]interface{}) (cursor accounting.BucketUsageCursor) {
 	limit, _ := args[LimitArg].(int)
 	page, _ := args[PageArg].(int)
