+++ satellite/satellitedb/consoledb.go
@@ -15,7 +15,7 @@ import (
 // ensures that ConsoleDB implements console.DB.
 var _ console.DB = (*ConsoleDB)(nil)
 
-// ConsoleDB contains access to different satellite databases
+// ConsoleDB contains access to different satellite databases.
 type ConsoleDB struct {
 	db *dbx.DB
 	tx *dbx.Tx
@@ -23,47 +23,42 @@ type ConsoleDB struct {
 	methods dbx.Methods
 }
 
-// Users is getter a for Users repository
+// Users is getter a for Users repository.
 func (db *ConsoleDB) Users() console.Users {
 	return &users{db.methods}
 }
 
-// Projects is a getter for Projects repository
+// Projects is a getter for Projects repository.
 func (db *ConsoleDB) Projects() console.Projects {
 	return &projects{db.methods}
 }
 
-// ProjectMembers is a getter for ProjectMembers repository
+// ProjectMembers is a getter for ProjectMembers repository.
 func (db *ConsoleDB) ProjectMembers() console.ProjectMembers {
 	return &projectMembers{db.methods, db.db}
 }
 
-// APIKeys is a getter for APIKeys repository
+// APIKeys is a getter for APIKeys repository.
 func (db *ConsoleDB) APIKeys() console.APIKeys {
 	return &apikeys{db.methods, db.db}
 }
 
-// RegistrationTokens is a getter for RegistrationTokens repository
+// RegistrationTokens is a getter for RegistrationTokens repository.
 func (db *ConsoleDB) RegistrationTokens() console.RegistrationTokens {
 	return &registrationTokens{db.methods}
 }
 
-// ResetPasswordTokens is a getter for ResetPasswordTokens repository
+// ResetPasswordTokens is a getter for ResetPasswordTokens repository.
 func (db *ConsoleDB) ResetPasswordTokens() console.ResetPasswordTokens {
 	return &resetPasswordTokens{db.methods}
 }
 
-// UsageRollups is a getter for console.UsageRollups repository
-func (db *ConsoleDB) UsageRollups() console.UsageRollups {
-	return &usagerollups{db.db}
-}
-
-// UserCredits is a getter for console.UserCredits repository
+// UserCredits is a getter for console.UserCredits repository.
 func (db *ConsoleDB) UserCredits() console.UserCredits {
 	return &usercredits{db.db, db.tx}
 }
 
-// BeginTx is a method for opening transaction
+// BeginTx is a method for opening transaction.
 func (db *ConsoleDB) BeginTx(ctx context.Context) (console.DBTx, error) {
 	if db.db == nil {
 		return nil, errs.New("DB is not initialized!")
@@ -84,12 +79,12 @@ func (db *ConsoleDB) BeginTx(ctx context.Context) (console.DBTx, error) {
 	}, nil
 }
 
-// DBTx extends Database with transaction scope
+// DBTx extends Database with transaction scope.
 type DBTx struct {
 	*ConsoleDB
 }
 
-// Commit is a method for committing and closing transaction
+// Commit is a method for committing and closing transaction.
 func (db *DBTx) Commit() error {
 	if db.tx == nil {
 		return errs.New("begin transaction before commit it!")
@@ -98,7 +93,7 @@ func (db *DBTx) Commit() error {
 	return db.tx.Commit()
 }
 
-// Rollback is a method for rollback and closing transaction
+// Rollback is a method for rollback and closing transaction.
 func (db *DBTx) Rollback() error {
 	if db.tx == nil {
 		return errs.New("begin transaction before rollback it!")
