+++ satellite/satellitedb/projects.go
@@ -34,6 +34,18 @@ func (projects *projects) GetAll(ctx context.Context) (_ []console.Project, err 
 	return projectsFromDbxSlice(ctx, projectsDbx)
 }
 
+// GetOwn is a method for querying all projects created by current user from the database.
+func (projects *projects) GetOwn(ctx context.Context, userID uuid.UUID) (_ []console.Project, err error) {
+	defer mon.Task()(&ctx)(&err)
+
+	projectsDbx, err := projects.db.All_Project_By_OwnerId_OrderBy_Asc_CreatedAt(ctx, dbx.Project_OwnerId(userID[:]))
+	if err != nil {
+		return nil, err
+	}
+
+	return projectsFromDbxSlice(ctx, projectsDbx)
+}
+
 // GetCreatedBefore retrieves all projects created before provided date
 func (projects *projects) GetCreatedBefore(ctx context.Context, before time.Time) (_ []console.Project, err error) {
 	defer mon.Task()(&ctx)(&err)
