+++ web/satellite/.env
@@ -1,2 +1,2 @@
-VUE_APP_STRIPE_PUBLIC_KEY=pk_test_
+VUE_APP_STRIPE_PUBLIC_KEY=pk_test_HMWSkLXmYBkQYeO9VRVQ8eHf00zrxjgwMh
 VUE_APP_ENDPOINT_URL=/api/v0/graphql
